Updating affirmation file structure, appearance and components
