+++ resources/assets/components/Affirmation/AffirmationContainer.js
@@ -1,0 +1,8 @@
+import { connect } from 'react-redux';
+import Affirmation from './Affirmation';
+
+const mapStateToProps = state => ({
+  content: state.campaign.affirmation,
+});
+
+export default connect(mapStateToProps)(Affirmation);
