+++ resources/assets/components/Affirmation/index.js
@@ -1,61 +1,3 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import { FlexCell } from '../Flex';
-import Highlight from '../Highlight';
-import { Figure } from '../Figure';
-import Wrapper from '../Wrapper';
-import ShareContainer from '../../containers/ShareContainer';
-import './affirmation.scss';
+export AffirmationContainer from './AffirmationContainer';
 
-const Affirmation = ({ content }) => (
-  <FlexCell width="full">
-    <div className="affirmation">
-      <Wrapper width="feed">
-        <div className="affirmation__section affirmation__section-heading">
-          <Highlight>{ content.header }</Highlight>
-        </div>
-        <div className="affirmation__section affirmation__section-quote">
-          <Figure image={content.photo} alt={content.author} alignment="left">
-            <p>{ content.quote }</p>
-            <span>- { content.author }</span>
-          </Figure>
-        </div>
-        <div className="affirmation__section affirmation__section-share">
-          <div className="affirmation__block">
-            <h3>{ content.callToActionHeader }</h3>
-            <p>{ content.callToActionDescription }</p>
-          </div>
-          <div className="affirmation__block">
-            <ShareContainer variant="blue" parentSource="affirmation" />
-          </div>
-        </div>
-      </Wrapper>
-    </div>
-  </FlexCell>
-);
-
-Affirmation.propTypes = {
-  content: PropTypes.shape({
-    header: PropTypes.string,
-    photo: PropTypes.string,
-    author: PropTypes.string,
-    quote: PropTypes.string,
-    callToActionHeader: PropTypes.string,
-    callToActionDescription: PropTypes.string,
-  }),
-};
-
-// @TODO: Only have this in place because not all campaigns have Affirmation content yet.
-// So delete this later when its safe.
-Affirmation.defaultProps = {
-  content: {
-    header: '',
-    photo: '',
-    author: '',
-    quote: '',
-    callToActionHeader: '',
-    callToActionDescription: '',
-  },
-};
-
-export default Affirmation;
+export default from './Affirmation';
