+++ resources/assets/components/Modal/modal.scss
@@ -22,8 +22,8 @@
       margin-top: 30%;
     }
 
-    // Override any child components
-    > div {
+    // Override block components
+    > .block-wrapper {
       margin: 0;
     }
   }
